редактирует модули
переносит содержимое модулей create-comment и create-photo в data.js;
переносит функцию postPhotos из main.js в data.js
